platform/opensens-common/pyproject.toml
[project]
name = "opensens-common"
version = "0.1.0"
description = "Shared utilities for OpenSens Darklab projects"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.39.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["opensens_common*"]
